Рефакторинг js кода
